seperate protocol ixs
