+++ boring-vault-svm-core/src/manage_instructions/jupiter.rs
@@ -1,0 +1,8 @@
+// TODO Jupiter DEX plan
+// Looks like an anchor program, so ingest discriminator
+// Do not ingest ix data size as that will change betweeen calls
+// Ingest destination token account, source mint, destination mint
+// example txs
+// https://solscan.io/tx/5LqV9oZUitaPieA3pth1JZoyo2XPcSfZawWioEtPUWAqTdRbDqqwy2Mjuw3cLM2mkcXeA4E5yAhFdhuYR9yYPtBz
+// https://solscan.io/tx/27fTvBLo2GuBLLL1jSEGKLqKtcioFwQQG6mNw4vcF54G9iE8FLBU1wQZWNZGkeqRocnf772Qx1DWSkyiQ37ocZx2
+// https://solscan.io/tx/4Cgipr3V7qgxmAFNv4ZqxowgBRxBiUtsGE4QC3gKav7FT3GZQUuB9R6z9KBms25SdvQd6pLEhhxphyaEmDnHMXaU
